Fixed gcc warnings, especially about dereferencing type-punned pointer
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@893 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip-apps/src/pjsua/pjsua_app.c
@@ -1114,7 +1114,7 @@ static int write_settings(const struct app_config *config,
 
     if (config->udp_cfg.tls_setting.timeout.sec) {
 	pj_ansi_sprintf(line, "--tls-neg-timeout %d\n",
-			config->udp_cfg.tls_setting.timeout.sec);
+			(int)config->udp_cfg.tls_setting.timeout.sec);
 	pj_strcat2(&cfg, line);
     }
